Fix clang-tidy warnings, add [[nodiscard]], update tooling
- Fix implicit widening in font.cpp (uint32_t * uint32_t -> size_t) and
  texture.cpp (uint32_t * uint32_t * 4 -> VkDeviceSize)
- Add default cases to GLFW action switches in window.cpp
- Remove redundant enumerate_entities() redeclaration in ISerializablePool
  that shadowed the IComponentPool virtual base method
- Add [[nodiscard]] across all public headers and internal vulkan headers
- Silence discarded-result warnings in examples/tests with (void) casts
- Update .clang-tidy: suppress known false positives (chained-comparison
  on template brackets, infinite-loop on side-effecting calls, Vulkan
  enum zero-init, identifier naming for public aggregate members)
- Update CLAUDE.md with clang-format/clang-tidy workflow instructions

+++ src/font.cpp
@@ -109,7 +109,7 @@ std::expected<BitmapFont, Error> BitmapFont::load_pcf(vk::Context& ctx,
     // Square-ish atlas: each glyph occupies one cell_w × cell_h slot.
     size_t num_glyphs = metrics.size();
     uint32_t cols = 1;
-    while (cols * cols < num_glyphs)
+    while (static_cast<size_t>(cols) * cols < num_glyphs)
         ++cols;
     uint32_t rows_atlas = (static_cast<uint32_t>(num_glyphs) + cols - 1) / cols;
     uint32_t atlas_w = cols * cell_w;
@@ -406,7 +406,7 @@ std::expected<BitmapFont, Error> BitmapFont::load_bdf(vk::Context& ctx,
 
     size_t num_glyphs = bf.glyphs.size();
     uint32_t cols = 1;
-    while (cols * cols < num_glyphs)
+    while (static_cast<size_t>(cols) * cols < num_glyphs)
         ++cols;
     uint32_t rows_atlas = (static_cast<uint32_t>(num_glyphs) + cols - 1) / cols;
     uint32_t atlas_w = cols * cell_w;
